refactor(server): Split server.go into console.go and s3api.go
