--- server/cors.go
@@ -1,22 +1,0 @@
-package server
-
-import "net/http"
-
-// corsHandler wraps next and adds permissive CORS headers to every response.
-// OPTIONS preflight requests are answered immediately with 200 OK.
-func corsHandler(next http.Handler) http.Handler {
-	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		h := w.Header()
-		h.Set("Access-Control-Allow-Origin", "*")
-		h.Set("Access-Control-Allow-Methods", "GET, PUT, POST, DELETE, HEAD, OPTIONS")
-		h.Set("Access-Control-Allow-Headers", "*")
-		h.Set("Access-Control-Max-Age", "86400")
-
-		if r.Method == http.MethodOptions {
-			w.WriteHeader(http.StatusOK)
-			return
-		}
-
-		next.ServeHTTP(w, r)
-	})
-}
